cheri: Fix warnings in kernel/
Mostly uintptr_t vs. unsigned long issues.

+++ kernel/trace/bpf_trace.c
@@ -2854,7 +2854,7 @@ kprobe_multi_link_prog_run(struct bpf_kprobe_multi_link *link,
 
 static int
 kprobe_multi_link_handler(struct fprobe *fp, unsigned long fentry_ip,
-			  unsigned long ret_ip, struct pt_regs *regs,
+			  uintptr_t ret_ip, struct pt_regs *regs,
 			  void *data)
 {
 	struct bpf_kprobe_multi_link *link;
@@ -2867,7 +2867,7 @@ kprobe_multi_link_handler(struct fprobe *fp, unsigned long fentry_ip,
 
 static void
 kprobe_multi_link_exit_handler(struct fprobe *fp, unsigned long fentry_ip,
-			       unsigned long ret_ip, struct pt_regs *regs,
+			       uintptr_t ret_ip, struct pt_regs *regs,
 			       void *data)
 {
 	struct bpf_kprobe_multi_link *link;
